% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate_species.R
\name{create_env_raster}
\alias{create_env_raster}
\title{Simulating environmental layer}
\usage{
create_env_raster(length = 2000)
}
\arguments{
\item{length}{length of the linear gradient (default = 2000)}
}
\description{
This function simulates a single linear gradient from a raster
}
