% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-predict_richness.R
\name{extract_rich_component}
\alias{extract_rich_component}
\title{Function to extract deviation of species richness}
\usage{
extract_rich_component(rich_df, component)
}
\description{
Function to extract deviation of species richness
}
