% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-predict_richness.R
\name{predict_prob_composition_accuracy}
\alias{predict_prob_composition_accuracy}
\title{Composition accuracy from predicted probabilities}
\usage{
predict_prob_composition_accuracy(sdm_df, n_draws = 10)
}
\description{
This function computes the accuracy (specificity and sensitivity) of the
prediction of composition of assemblages. It quantifies the amount to which
the predicted probabilities match well the real presence and the real
absence of species at the assemblage level. To do so it needs to draw
distribution from the probabilities and computes the average specificity and
sensitivity at the assemblage level
}
