% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-predict_richness.R
\name{predict_thresh_composition_accuracy}
\alias{predict_thresh_composition_accuracy}
\title{Composition accuracy from threshold predictions}
\usage{
predict_thresh_composition_accuracy(sdm_df)
}
\description{
This function computes the accuracy (specificity and sensitivity) of the
prediction of composition of assemblages. It quantifies the amount to which
the prediction using thresholds predict well the real presence and the real
absence of species at the assemblage level
}
