% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate_species.R
\name{sample_species_pa}
\alias{sample_species_pa}
\title{Randomly sample species presence-absence}
\usage{
sample_species_pa(species_list, virt_raster, n_cell = 300)
}
\arguments{
\item{species_list}{a list of virtual species produced by the
\code{virtualspecies} package}

\item{n_cell}{number of cells to sample}
}
\description{
This function sample presences absences from a virtual species list on a
given number of sites. It outputs a data.frame containing presences and
absences of species all in the same sites.
}
