% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-simulate_species.R
\name{simulate_species}
\alias{simulate_species}
\title{Simulate Random Species with gaussian responses}
\usage{
simulate_species(n = 100, env_raster, given_seed = 27022018)
}
\arguments{
\item{n}{number of species to simulate}

\item{env_raster}{environmental raster}
}
\description{
Wrapper around the function \code{generateRandomSp()} of \code{virtualspecies}. This
function generates species with gaussian environmental response, with random
response curve parameters
}
